+++ Library/src/main/java/info/metadude/android/typedpreferences/FloatPreference.java
@@ -5,10 +5,8 @@ import android.content.SharedPreferences;
 /**
  * A wrapper class for a float preference.
  */
-public class FloatPreference {
+public class FloatPreference extends BasePreference {
 
-    protected final SharedPreferences mPreferences;
-    protected final String mKey;
     protected final float mDefaultValue;
     public static final float DEFAULT_VALUE_VALUE = 0f;
 
@@ -25,8 +23,7 @@ public class FloatPreference {
      * having the default value available.
      */
     public FloatPreference(final SharedPreferences preferences, final String key, float defaultValue) {
-        mPreferences = preferences;
-        mKey = key;
+        super(preferences, key);
         mDefaultValue = defaultValue;
     }
 
@@ -39,25 +36,10 @@ public class FloatPreference {
     }
 
     /**
-     * Returns {@code true} if some value is stored for
-     * this preference, otherwise {@code false}.
-     */
-    public boolean isSet() {
-        return mPreferences.contains(mKey);
-    }
-
-    /**
      * Stores the given {@code float} value asynchronously.
      */
     public void set(float value) {
         mPreferences.edit().putFloat(mKey, value).apply();
-    }
-
-    /**
-     * Removes this preference setting asynchronously.
-     */
-    public void delete() {
-        mPreferences.edit().remove(mKey).apply();
     }
 
 }
