+++ Library/src/main/java/info/metadude/android/typedpreferences/StringPreference.java
@@ -5,10 +5,8 @@ import android.content.SharedPreferences;
 /**
  * A wrapper class for a String preference.
  */
-public class StringPreference {
+public class StringPreference extends BasePreference {
 
-    protected final SharedPreferences mPreferences;
-    protected final String mKey;
     protected final String mDefaultValue;
     public static final String DEFAULT_VALUE_VALUE = "";
 
@@ -25,8 +23,7 @@ public class StringPreference {
      * having the default value available.
      */
     public StringPreference(final SharedPreferences preferences, final String key, final String defaultValue) {
-        mPreferences = preferences;
-        mKey = key;
+        super(preferences, key);
         mDefaultValue = defaultValue;
     }
 
@@ -39,25 +36,10 @@ public class StringPreference {
     }
 
     /**
-     * Returns {@code true} if some value is stored for
-     * this preference, otherwise {@code false}.
-     */
-    public boolean isSet() {
-        return mPreferences.contains(mKey);
-    }
-
-    /**
      * Stores the given {@code String} value asynchronously.
      */
     public void set(final String value) {
         mPreferences.edit().putString(mKey, value).apply();
-    }
-
-    /**
-     * Removes this preference setting asynchronously.
-     */
-    public void delete() {
-        mPreferences.edit().remove(mKey).apply();
     }
 
 }
